feat: CI - Make path-filter in `version.json` redundant (#5167)
## Description

Given that we're not using the path filter in any active capacity and
that it theoretically allows any contributor to influence what triggers
a new module version to be published this PR suggests to remove the same
at least until we make any use of the feature. Right now it's almost
'dead code' in all these version files.

- [ ] TODO: Create PR for
[schema](https://raw.githubusercontent.com/Azure/bicep/main/src/Bicep.RegistryModuleTool/JsonSchemas/schema.version.json)
to be merged AFTER this PR goes in (- the other way around would provoke
an error for everybody)

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|          |

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- [x] Update to CI Environment or utilities (Non-module affecting
changes)
- [ ] Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

---------

Co-authored-by: Benjamin P. Jung <b.jung@qvest-digital.com>
Co-authored-by: Benjamin P. Jung <headcr4sh@gmail.com>
Co-authored-by: Eric Scheffler <70648471+ericscheffler@users.noreply.github.com>
Co-authored-by: René Hézser <rene@hezser.de>
Co-authored-by: Erika Gressi <erikag@microsoft.com>