feat: Ptn ai foundry (#5475)
## Description

This is the **initial submission** of the AI-Foundry pattern module.
This will deploy either Basic (AI Foundry Account and project-
leveraging the updated Cognitive Services / Account AVM module) or
deploy the Standard Agent Private - creating the AI Foundry Account,
Project, Capability Host, CosmosDB, Storage and AI Search along with a
vNet, jump box and Bastion. Private Endpoints are also added and
configured. The Vnet has the Foundry Agent Service subnet added as well.
The feature flag will be expanded to include Standard Public in the next
iteration.
This will also address issue #5390
https://github.com/Azure/bicep-registry-modules/issues/5390 Which
requested the project and project connections to be added to the
cognitive services/account module. We should not add children to the
resource module for cognitive, as it may not be backward compatable.
Using this new pattern module for AI Foundry will apply all the
necessary parameters to the cognitive services/account to make it the AI
Foundry Account and then adds project and other Agent services.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->


[![avm.ptn.ai-ml.ai-foundry](https://github.com/mswantek68/bicep-registry-modules/actions/workflows/avm.ptn.ai-ml.ai-foundry.yml/badge.svg)](https://github.com/mswantek68/bicep-registry-modules/actions/workflows/avm.ptn.ai-ml.ai-foundry.yml)

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [ ] Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings

<!-- Please keep up to date with the contribution guide at
https://aka.ms/avm/contribute/bicep -->

---------

Co-authored-by: Mike Swantek <mike.swantek@microsoft.com>
Co-authored-by: Sebastian Gräf <sebastian@graef.io>
Co-authored-by: Seth Steenken <sethsteenken@gmail.com>
Co-authored-by: Seth <sethsteenken@microsoft.com>