feat: New Pattern Module `avm/ptn/sa/content processing (#5498)
## Description
This module contains the resources required to deploy the [Content
Processing solution
accelerator](https://github.com/microsoft/content-processing-solution-accelerator)
for both Sandbox environments and WAF aligned environments.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|
[![avm.ptn.sa.content-processing](https://github.com/brittneek/bicep-registry-modules/actions/workflows/avm.ptn.sa.content-processing.yml/badge.svg)](https://github.com/brittneek/bicep-registry-modules/actions/workflows/avm.ptn.sa.content-processing.yml)
|

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [x] Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings

##Notes

## Skipped WAF-Aligned PSRules

I skipped a few of the WAF-aligned PSRules.
https://github.com/brittneek/bicep-registry-modules/actions/runs/15836903820.
Below are the rules and the reasons for skipping them:

| Rule Name | Synopsis | Reason |

|--------------------------------------|-----------------------------------------------------------------------------------------------|--------|
| `Azure.ContainerApp.Insecure` | Ensure insecure inbound traffic is not
permitted to the container app. | The container apps are meant to allow
all public access for the accelerator to be usable. |
| `Azure.ContainerApp.RestrictIngress` | IP ingress restrictions mode
should be set to allow action for all rules defined. | The container
apps are meant to allow all public access for the accelerator to be
usable. Also, there's a circular reference with the frontend and backend
apps that prevents setting a CORS policy due to both apps needing each
other's FQDN. |
| `Azure.ContainerApp.PublicAccess` | Ensure public network access for
Container Apps environment is disabled. | The container apps are meant
to allow all public access for the accelerator to be usable. |
| `Azure.Cosmos.DisableMetadataWrite` | Use Entra ID identities for
management plane operations in Azure Cosmos DB. | Not applicable to this
accelerator pattern. |
| `Azure.Log.Replication` | Log Analytics workspaces should have
workspace replication enabled to improve availability. | The AVM module
has not been updated to support this yet. |
| `Azure.VNET.SubnetNaming` | Virtual Network subnets without a standard
naming convention may be difficult to manage. | The subnet `vnet-cps`
does not follow standard naming, but is acceptable for this sandboxed
accelerator. |
| *(Default Configuration Reliability)*| PSRules require enabling
certain Reliability features in the default config. | This pattern is a
closed accelerator to showcase AI scenarios and deploys a sandbox
configuration by default. These Reliability features are not needed. |

### Additional Notes

- When deploying the App Config AVM module, we receive an error that a
parameter value is null or not defined. This is incorrect because the
values are being set.
https://github.com/Azure/PSRule.Rules.Azure/issues/3434

### E2E Tests Skipped for PSRules

The following end-to-end tests were left in this version to skip the
PSRules:

- `tests/e2e/sandbox`
- `tests/e2e/enterprise-grade`

---------

Co-authored-by: DB Lee <donlee@microsoft.com>
Co-authored-by: Alvaro Guadamillas Herranz <alguadam@microsoft.com>