feat: New Pattern Module `avm/ptn/sa/modernize-your-code` (#5471)
## Description

This module contains the resources required to deploy the [Modernize
Your Code solution
accelerator](https://github.com/microsoft/Modernize-your-code-solution-accelerator)
for both Sandbox environments and WAF aligned environments.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|
[![avm.ptn.sa.modernize-your-code](https://github.com/sethsteenken/bicep-registry-modules-3/actions/workflows/avm.ptn.sa.modernize-your-code.yml/badge.svg)](https://github.com/sethsteenken/bicep-registry-modules-3/actions/workflows/avm.ptn.sa.modernize-your-code.yml)
|

## Type of Change

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [x] Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings

## Notes

I skipped a few of the waf-aligned PSRules. [View the validation results
here](https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/15912805423).
They are listed here with reasons for skipping:

| Rule Name | Reason |

|----------------------------------|----------------------------------------------------------------|
|
[Azure.ContainerApp.RestrictIngress](https://azure.github.io/PSRule.Rules.Azure/en/rules/Azure.ContainerApp.RestrictIngress)
| The container apps are meant to allow all public access for the
accelerator to be usable. Also there's a circular reference with the
frontend and backend apps that prevent setting a CORS policy due to both
apps needing each other's FQDN |
|
[Azure.ContainerApp.PublicAccess](https://azure.github.io/PSRule.Rules.Azure/en/rules/Azure.ContainerApp.PublicAccess)
| The container apps are meant to allow all public access for the
accelerator to be usable.
|
[Azure.Log.Replication](https://azure.github.io/PSRule.Rules.Azure/en/rules/Azure.Log.Replication)
| The AVM module has not been updated to support this yet |

This pattern (which in reality is not an open, re-usable pattern but a
closed accelerator to showcase AI scenarios) will deploy as default
version a sandbox configuration of the solution. PSRules require that
default configuration enables certain Reliability features that we don't
need to include.

The e2e tests that I left for this version to skip the PSRules are:
tests/e2e/sandbox and tests/e2e/enterprise-grade.

---------

Co-authored-by: Seth <sethsteenken@microsoft.com>
Co-authored-by: Alvaro Guadamillas Herranz <alvaroguadamillas@gmail.com>
Co-authored-by: René Hézser <rene@hezser.de>