feat: `avm/ptn/ai-ml/ai-foundry` - capability host (#5801)
## Description

- Added support for [Capability
Hosts](https://learn.microsoft.com/en-us/azure/ai-foundry/agents/concepts/capability-hosts)
for Account and Project
- Added support for `networkInjections`. Details
[here](https://learn.microsoft.com/en-us/azure/templates/microsoft.cognitiveservices/accounts?pivots=deployment-language-arm-template#networkinjections-1).
- Updated all `privateDnsZoneId` references to
`privateDnsZoneResourceId`
- Updated all `subnetId` references to `subnetResourceId`
- Updated account networking to line up with Terraform version

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
|----------|
|
[![avm.ptn.ai-ml.ai-foundry](https://github.com/sethsteenken/bicep-registry-modules-3/actions/workflows/avm.ptn.ai-ml.ai-foundry.yml/badge.svg?branch=avm/ptn/ai-ml/ai-foundry/version-030)](https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/17041661454)
|


## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [x] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation
- [ ] Update to CI Environment or utilities (Non-module affecting
changes)

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings
- [x] I have updated the module's CHANGELOG.md file with an entry for
the next version

## Notes

- ~~Currently implementing a workaround for test failures due to
Cognitive Services Account resource having a lock on the agent subnet
when doing `networkInjections`. This causes the deployment removal
process to fail (after a successful deployment). This involves a
post-deployment deploymentScript to execute some manual deletion tasks.
This seems a little messy, but should eventually work.~~
- ~~Will update CHANGELOG~~
- ~~Will update `version.json` to 0.3~~

## Test Ignores

Currently ignoring the `max` and `waf-aligned` tests due to issues with
the resource removal step. These two tests include a VNET and inject a
subnet in the Cognitive Services Account resource to be used for
"agents". The resource creates a link to the subnet but seems to lock
down this link causing errors when trying to remove resources.

```console
Subnet agents is in use by /subscriptions/***/resourceGroups/dep-***-bicep-fndrymax-rg/providers/Microsoft.Network/virtualNetworks/vnet-***fndrymax/subnets/agents/serviceAssociationLinks/legionservicelink and cannot be deleted. In order to delete the subnet, delete all the resources within the subnet.
```

Attempts were made to work around this - see [module for testing
here](https://github.com/sethsteenken/bicep-registry-modules-3/blob/avm/ptn/ai-ml/ai-foundry/version-030/avm/ptn/ai-ml/ai-foundry/tests/shared/removeLockingDependencies.bicep)
that would execute a `deploymentScript` to force a disconnect of the
Cognitive Services account the subnet. The results from this script were
too inconsistent and because the tests take almost 2 hours to complete,
the time to debug the script issues was not favorable.

On the Terraform version of this pattern module, the [test simply waits
for 10 mins and attempts to perform a purge of the Cognitive Services
account
resource](https://github.com/Azure/terraform-azurerm-avm-ptn-aiml-ai-foundry/blob/main/examples/standard-private/main.tf#L386)
assuming it's been deleted by then. This is not really possible in Bicep
and frankly, should not be the approach we take to work-around a
resource dependency issue.

To show the tests are successful at deploying the resources, here are
completed workflows that successfully deployed the `max` and
`waf-aligned` tests but failed to remove them:

-
[https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/16995958427](https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/16995958427)
-
[https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/16993512773](https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/16993512773)
-
[https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/16975571024](https://github.com/sethsteenken/bicep-registry-modules-3/actions/runs/16975571024)

---------

Co-authored-by: sethsteenken <sethsteenken@microsoft.com>
Co-authored-by: Mike Swantek <46489667+mswantek68@users.noreply.github.com>