serverFarmResourceId can be '', adjust some related param validation
The serverFarmResourceId is '' when deploying to a containerapp.
However, there were a few parameters (clientAffinityEnabled,
clientCertMode, publicNetworkAccess) that no longer apply and should be
set to null in these cases. Otherwise, they prevent the deployment of
these resources. These changes fix the behavior by setting these to null
if the serverFarmResourceId is ''.