fix: Some Bicep issues when trying to deploy ContainerApps (#5898)
## Description

Fixes a couple of incompatibilities when attempting to use the
`web/sites` module for deploying a function app in a containerapp
environment. The current implementation allows `serverFarmResourceId` to
be `''` in order to support ContainerApp rather than App Service Plan
deployment. However, this means `clientAffinityEnabled`,
`clientCertMode`, and `publicNetworkAccess` should all be `null` as
these appear to be handled by the Container App deployment. This module
currently attempts to set these to a defined value, which is
incompatible and results in a failed deployment.

Similarly, while `customDomainVerificationId` is defined as optional
(can be `null`) on input, the output type requires it to be a `string`.
These changes also fix this by making the output allowed to be `null` as
well.

## Pipeline Status


[![avm.res.web.site](https://github.com/ckane/bicep-registry-modules/actions/workflows/avm.res.web.site.yml/badge.svg?branch=containerapp-issues-ckane)](https://github.com/ckane/bicep-registry-modules/actions/workflows/avm.res.web.site.yml)

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- Azure Verified Module updates:
- [x] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation
- [ ] Update to CI Environment or utilities (Non-module affecting
changes)

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings
- [x] I have updated the module's CHANGELOG.md file with an entry for
the next version

<!-- Please keep up to date with the contribution guide at
https://aka.ms/avm/contribute/bicep -->