feat: Add support for standalone Network Security Group (NSG) in sub-vending (#5970)
## Description

I am currently implementing sub-vending with the requirement to deploy a
"clean" Azure Landing Zone. In this case, clean means a subscription, a
virtual network with peering to the hub (no subnets), a route table and
a network security group. Deploying an NSG was not possible, so I added
an optional parameter to be able to deploy it standalone instead via the
subnets parameter.

Tagging module owners: @jtracey93 @sebassem 

What is the way to test? I believe I am not able to test this in my MVP
tenant, and I don't have another environment to deploy subscriptions.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

(static tests)
| Pipeline |
| -------- |

[![avm.ptn.lz.sub-vending](https://github.com/johnlokerse/bicep-registry-modules/actions/workflows/avm.ptn.lz.sub-vending.yml/badge.svg?branch=johnlokerse%2Ffix-standalone-nsg-sub-vending)](https://github.com/johnlokerse/bicep-registry-modules/actions/workflows/avm.ptn.lz.sub-vending.yml)


## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- Azure Verified Module updates:
- [x] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [x] Update to documentation
- [ ] Update to CI Environment or utilities (Non-module affecting
changes)

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [ ] My corresponding pipelines / checks run clean and green without
any errors or warnings
- [x] I have updated the module's CHANGELOG.md file with an entry for
the next version

<!-- Please keep up to date with the contribution guide at
https://aka.ms/avm/contribute/bicep -->