ci: Retire Contributor teams for Bicep module (#6082)
This PR is to change the automation scripts which will remove reference
of any teams for contributors.

There is no reference to any issues for this PR since this is related
contributor teams not being used and also removing the display of
Contributor teams on AVM website.

## Pipeline Reference
Nil

## Type of Change
This is change in the script which reads the CSV file so removed the
reference to the contributor teams as these contributor related data
will not be part of CSV going forward.

- Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation
- [ ] Update to CI Environment or utilities (Non-module affecting
changes)

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [ ] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [ ] My corresponding pipelines / checks run clean and green without
any errors or warnings
- [ ] I have updated the module's CHANGELOG.md file with an entry for
the next version

---------

Co-authored-by: AMNIRANT <amnirant@microsoft.com>