fix: Update cdn/profile secretVersion to handle optional values (#6097)
## Description

If secretVersion is not passed in the type, it fails with error:

``` text
'The language expression property 'secretVersion' doesn't exist, available properties are 'name, type, useLatestVersion, secretSourceResourceId'.'
```

By checking for null in the optional value and defaulting to '' we no
longer are required to pass the secretVersion property.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |


## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- Azure Verified Module updates:
- [X] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation
- [ ] Update to CI Environment or utilities (Non-module affecting
changes)

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` (`bicep build`) locally to generate the
supporting module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings
- [x] I have updated the module's CHANGELOG.md file with an entry for
the next version

<!-- Please keep up to date with the contribution guide at
https://aka.ms/avm/contribute/bicep -->